+++ src/components/TableForPage/TableForPage.tsx
@@ -1,0 +1,59 @@
+import classNames from 'classnames';
+import { PersonLink } from '../PersonLink/PersonLink';
+import { Person } from '../../types';
+import React from 'react';
+
+type Props = {
+  people: Person[];
+  personId?: string;
+};
+
+export const TableForPage: React.FC<Props> = ({ people, personId }) => {
+  return (
+    <table
+      data-cy="peopleTable"
+      className="table is-striped is-hoverable is-narrow is-fullwidth"
+    >
+      <thead>
+        <tr>
+          <th>Name</th>
+          <th>Sex</th>
+          <th>Born</th>
+          <th>Died</th>
+          <th>Mother</th>
+          <th>Father</th>
+        </tr>
+      </thead>
+
+      <tbody>
+        {people.map(person => {
+          const mother = people.find(p => p.name === person.motherName);
+          const father = people.find(p => p.name === person.fatherName);
+
+          return (
+            <tr
+              data-cy="person"
+              key={person.slug}
+              className={classNames({
+                'has-background-warning': person.slug === personId,
+              })}
+            >
+              <td>
+                <PersonLink person={person} name={person.name} />
+              </td>
+              <td>{person.sex}</td>
+              <td>{person.born}</td>
+              <td>{person.died}</td>
+              <td>
+                <PersonLink person={mother} name={person.motherName} />
+              </td>
+              <td>
+                <PersonLink person={father} name={person.fatherName} />
+              </td>
+            </tr>
+          );
+        })}
+      </tbody>
+    </table>
+  );
+};
