+++ src/components/TableForPage/index.ts
@@ -1,0 +1,1 @@
+export { TableForPage } from './TableForPage';
