+++ src/components/TablePage/TablePage.tsx
@@ -2,9 +2,8 @@ import { useEffect, useState } from 'react';
 import { Loader } from '../Loader';
 import { getPeople } from '../../api';
 import { Person } from '../../types';
-import { PersonLink } from '../PersonLink/PersonLink';
 import { useParams } from 'react-router-dom';
-import classNames from 'classnames';
+import { TableForPage } from '../TableForPage';
 
 export const TablePage = () => {
   const [people, setPeople] = useState<Person[]>([]);
@@ -47,51 +46,7 @@ export const TablePage = () => {
         )}
 
         {people.length > 0 && (
-          <table
-            data-cy="peopleTable"
-            className="table is-striped is-hoverable is-narrow is-fullwidth"
-          >
-            <thead>
-              <tr>
-                <th>Name</th>
-                <th>Sex</th>
-                <th>Born</th>
-                <th>Died</th>
-                <th>Mother</th>
-                <th>Father</th>
-              </tr>
-            </thead>
-
-            <tbody>
-              {people.map(person => {
-                const mother = people.find(p => p.name === person.motherName);
-                const father = people.find(p => p.name === person.fatherName);
-
-                return (
-                  <tr
-                    data-cy="person"
-                    key={person.slug}
-                    className={classNames({
-                      'has-background-warning': person.slug === personId,
-                    })}
-                  >
-                    <td>
-                      <PersonLink person={person} name={person.name} />
-                    </td>
-                    <td>{person.sex}</td>
-                    <td>{person.born}</td>
-                    <td>{person.died}</td>
-                    <td>
-                      <PersonLink person={mother} name={person.motherName} />
-                    </td>
-                    <td>
-                      <PersonLink person={father} name={person.fatherName} />
-                    </td>
-                  </tr>
-                );
-              })}
-            </tbody>
-          </table>
+          <TableForPage people={people} personId={personId} />
         )}
       </div>
     </div>
